Backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.15
pg8000==1.30.3
pandas==2.0.3
python-multipart==0.0.6
alembic==1.11.3
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
